+++ pkg/node/node.go
@@ -60,6 +60,7 @@ type Node struct {
 	ctx            context.Context
 	info           BuildInfo
 	configCh       chan WunderNodeConfig
+	builder        *apihandler.Builder
 	server         *http.Server
 	pool           *pool.Pool
 	log            *zap.Logger
@@ -248,8 +249,17 @@ func (n *Node) Shutdown(ctx context.Context) error {
 }
 
 func (n *Node) Close() error {
+	if n.builder != nil {
+		if err := n.builder.Close(); err != nil {
+			return err
+		}
+		n.builder = nil
+	}
 	if n.server != nil {
-		return n.server.Close()
+		if err := n.server.Close(); err != nil {
+			return err
+		}
+		n.server = nil
 	}
 	return nil
 }
@@ -264,7 +274,7 @@ func (n *Node) newListeners(configuration *apihandler.Listener) ([]net.Listener,
 	var listeners []net.Listener
 	var localhostIPs []net.IP
 
-	// If listening to 'localhost', listen to both 127.0.0.1 or ::1 if they are available.
+	// If listening to 'localhost', listen to both localhost or ::1 if they are available.
 	if host == "localhost" {
 		localhostIPs, _ = net.LookupIP(host)
 	}
@@ -273,6 +283,10 @@ func (n *Node) newListeners(configuration *apihandler.Listener) ([]net.Listener,
 		nip := ip.String()
 		// isIPv6
 		if strings.Contains(nip, ":") {
+			// Filter out link-local addresses
+			if strings.HasPrefix(nip, "fe80:") {
+				continue
+			}
 			listener, err := cfg.Listen(context.Background(), "tcp6", fmt.Sprintf("[%s]:%d", nip, port))
 			// in some cases e.g when ipv6 is not enabled in docker, listen will error
 			// in that case we ignore this error and try to listen to next ip
@@ -435,10 +449,10 @@ func (n *Node) startServer(nodeConfig WunderNodeConfig) error {
 		DevMode:                    n.options.devMode,
 	}
 
-	builder := apihandler.NewBuilder(n.pool, n.log, loader, hooksClient, builderConfig)
+	n.builder = apihandler.NewBuilder(n.pool, n.log, loader, hooksClient, builderConfig)
 	internalBuilder := apihandler.NewInternalBuilder(n.pool, n.log, loader)
 
-	publicClosers, err := builder.BuildAndMountApiHandler(n.ctx, router, nodeConfig.Api)
+	publicClosers, err := n.builder.BuildAndMountApiHandler(n.ctx, router, nodeConfig.Api)
 	if err != nil {
 		n.log.Error("BuildAndMountApiHandler", zap.Error(err))
 	}