+++ pkg/node/node_test.go
@@ -8,6 +8,7 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"io"
 	"net/http"
 	"net/http/httptest"
 	"net/http/httputil"
@@ -126,7 +127,7 @@ func TestNode(t *testing.T) {
 			},
 			Options: &apihandler.Options{
 				Listener: &apihandler.Listener{
-					Host: "127.0.0.1",
+					Host: "localhost",
 					Port: uint16(port),
 				},
 				Logging: apihandler.Logging{Level: zap.ErrorLevel},
@@ -199,6 +200,127 @@ func TestNode(t *testing.T) {
 		httpexpect.ContentOpts{MediaType: "text/html"})
 }
 
+func TestInMemoryCache(t *testing.T) {
+	logger := logging.New(true, false, zapcore.DebugLevel)
+
+	productService := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		data, err := io.ReadAll(r.Body)
+		if err != nil {
+			t.Fatal(err)
+		}
+		if bytes.Contains(data, []byte(`{"variables":{},"query":"query($first: Int){topProducts(first: $first){upc name price}}"}`)) {
+			if _, err := io.WriteString(w, `{"data":{"topProducts":[{"upc":"1","name":"A","price":1},{"upc":"2","name":"B","price":2}]}}`); err != nil {
+				t.Fatal(err)
+			}
+			return
+		}
+		w.WriteHeader(500)
+	}))
+	defer productService.Close()
+
+	port, err := freeport.GetFreePort()
+	assert.NoError(t, err)
+
+	nodeURL := fmt.Sprintf(":%d", port)
+
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	node := New(ctx, BuildInfo{}, "", logger)
+
+	nodeConfig := WunderNodeConfig{
+		Server: &Server{
+			GracefulShutdownTimeout: 0,
+			KeepAlive:               5,
+			ReadTimeout:             5,
+			WriteTimeout:            5,
+			IdleTimeout:             5,
+		},
+		Api: &apihandler.Api{
+			Hosts:                 []string{"jens.wundergraph.dev"},
+			EngineConfiguration:   federationPlanConfiguration("", productService.URL, ""),
+			EnableSingleFlight:    true,
+			EnableGraphqlEndpoint: true,
+			Operations: []*wgpb.Operation{
+				{
+					Name:    "TopProducts",
+					Content: topProductsQuery,
+					CacheConfig: &wgpb.OperationCacheConfig{
+						Enable:               true,
+						MaxAge:               60,
+						Public:               true,
+						StaleWhileRevalidate: 60,
+					},
+					OperationType: wgpb.OperationType_QUERY,
+					HooksConfiguration: &wgpb.OperationHooksConfiguration{
+						MockResolve: &wgpb.MockResolveHookConfiguration{},
+					},
+					VariablesSchema:              `{"type":"object","properties":{"first":{"type":["number","null"]}}}`,
+					ResponseSchema:               `{}`,
+					InterpolationVariablesSchema: `{"type":"object","properties":{"first":{"type":["number","null"]}}}`,
+					AuthorizationConfig: &wgpb.OperationAuthorizationConfig{
+						RoleConfig: &wgpb.OperationRoleConfig{},
+					},
+				},
+			},
+			AuthenticationConfig: &wgpb.ApiAuthenticationConfig{
+				CookieBased: &wgpb.CookieBasedAuthentication{},
+				JwksBased:   &wgpb.JwksBasedAuthentication{},
+				Hooks:       &wgpb.ApiAuthenticationHooks{},
+			},
+			CacheConfig: &wgpb.ApiCacheConfig{
+				Kind: wgpb.ApiCacheKind_IN_MEMORY_CACHE,
+				InMemoryConfig: &wgpb.InMemoryCacheConfig{
+					MaxSize: 16 * 1024 * 1024,
+				},
+			},
+			Options: &apihandler.Options{
+				Listener: &apihandler.Listener{
+					Host: "localhost",
+					Port: uint16(port),
+				},
+				Logging: apihandler.Logging{Level: zap.ErrorLevel},
+			},
+		},
+	}
+
+	go func() {
+		err = node.StartBlocking(WithStaticWunderNodeConfig(nodeConfig))
+		assert.NoError(t, err)
+	}()
+
+	time.Sleep(time.Second)
+
+	e := httpexpect.WithConfig(httpexpect.Config{
+		BaseURL: "http://" + nodeURL,
+		Client: &http.Client{
+			Jar:     httpexpect.NewJar(),
+			Timeout: time.Second * 30,
+		},
+		Reporter: httpexpect.NewRequireReporter(t),
+	})
+
+	withHeaders := e.Builder(func(request *httpexpect.Request) {
+		request.WithHeader("Host", "jens.wundergraph.dev")
+	})
+
+	const (
+		cacheHeaderName  = "X-Wg-Cache"
+		expectedDataName = "top products without query"
+	)
+
+	// Send a request to populate the cache
+	cold := withHeaders.GET("/operations/TopProducts").Expect()
+	cold.Status(http.StatusOK).Header(cacheHeaderName).Equal("MISS")
+	goldie.Assert(t, expectedDataName, prettyJSON(cold.Body().Raw()))
+
+	// Close the origin, so request can only be answered from the in-memory cache
+	productService.Close()
+	hot := withHeaders.GET("/operations/TopProducts").Expect()
+	hot.Status(http.StatusOK).Header(cacheHeaderName).Equal("HIT")
+	goldie.Assert(t, expectedDataName, prettyJSON(hot.Body().Raw()))
+}
+
 func TestWebHooks(t *testing.T) {
 
 	var paths []string
@@ -261,7 +383,7 @@ func TestWebHooks(t *testing.T) {
 			Options: &apihandler.Options{
 				ServerUrl: testServer.URL,
 				Listener: &apihandler.Listener{
-					Host: "127.0.0.1",
+					Host: "localhost",
 					Port: uint16(port),
 				},
 				Logging: apihandler.Logging{Level: zap.ErrorLevel},
@@ -357,7 +479,7 @@ func BenchmarkNode(t *testing.B) {
 			},
 			Options: &apihandler.Options{
 				Listener: &apihandler.Listener{
-					Host: "127.0.0.1",
+					Host: "localhost",
 					Port: uint16(port),
 				},
 				Logging: apihandler.Logging{Level: zap.ErrorLevel},