+++ packages/nextjs/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.7.0...@wundergraph/nextjs@0.7.1) (2022-12-15)
+
+**Note:** Version bump only for package @wundergraph/nextjs
+
+## [0.7.0](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.6.1...@wundergraph/nextjs@0.7.0) (2022-12-14)
+
+### Features
+
+* update SWR to v2 🥳 ([#446](https://github.com/wundergraph/wundergraph/issues/446)) ([8af6b99](https://github.com/wundergraph/wundergraph/commit/8af6b99f9e0e9da84586782ad5b9ad8e1072c987)) (@Pagebakers)
+
+### Bug Fixes
+
+* add missing dependency to our nextjs package ([#432](https://github.com/wundergraph/wundergraph/issues/432)) ([d646d93](https://github.com/wundergraph/wundergraph/commit/d646d933bcc1a67301446fa44b4d27c934ab94da)) (@fiam)
+
 ## [0.6.1](https://github.com/wundergraph/wundergraph/compare/@wundergraph/nextjs@0.6.0...@wundergraph/nextjs@0.6.1) (2022-12-08)
 
 **Note:** Version bump only for package @wundergraph/nextjs